Moved filter_attribute tests
